feat: split fixture value equally among team members
Players on a 2-member team now receive/lose half the fixture value
rather than the full amount. Solo teams are unaffected.

- Add migration 000006: recreate user_total_score view to divide
  f.value by the team's current member_count (NULLIF-guarded)
- Change entity.User.TotalScore from int to float64 to accommodate
  fractional scores when an odd value is split across a 2-player team
- Simplify GetByUsername to join user_total_score view for score
  rather than maintaining a separate inline query
- Update openapi.yaml: totalScore type integer → number (format: double)

+++ internal/entity/user.go
@@ -12,10 +12,10 @@ const (
 // hashsalt) and the force_password_change flag are unexported so they are
 // never accidentally serialised in HTTP responses.
 type User struct {
-	Id         int    `json:"id"`
-	Username   string `json:"username"`
-	Role       Role   `json:"role"`
-	TotalScore int    `json:"totalScore"`
+	Id         int     `json:"id"`
+	Username   string  `json:"username"`
+	Role       Role    `json:"role"`
+	TotalScore float64 `json:"totalScore"`
 	password   string
 	hashsalt   string
 	// ForcePasswordChange signals that the user must change their password
@@ -31,7 +31,7 @@ func (u User) GetHashsalt() string { return u.hashsalt }
 // NewUser constructs a fully-populated User including the credential fields.
 // Intended for use by the repository layer when loading rows that include
 // password, hashsalt, and force_password_change (e.g. for authentication).
-func NewUser(id int, username string, role Role, password, hashsalt string, totalScore int, forcePasswordChange bool) User {
+func NewUser(id int, username string, role Role, password, hashsalt string, totalScore float64, forcePasswordChange bool) User {
 	return User{
 		Id:                  id,
 		Username:            username,
